Fix jpa tests
The tables persists after the tests for verification and visualization
of the data in an actual test database.
The create option of spring.jpa.hibernate.ddl-auto isn't documented in
the spring docs https://docs.spring.io/spring-boot/docs/1.1.0.M1/reference/html/howto-database-initialization.html
This initializes the database dropping the tables before create the
inverse of create-drop option, allowing the database to persist after
the test and cleaning before the test.