Fix h2 username password incorrect
username and password in h2 using persistence database needs to be
defined in spring.datasource.username and spring.datasource.password
for you to login to avoid confusion.
Because if not defined h2 by default sets the username and the password
by default, however sets differently for in-memory and persistent mode.
in-memory mode username is set to "sa" and password is empty.
in persistence mode username and password they're empty.